Add a overview tab on fellowship member profile page, #7463
